Force Elements

Bushing

Modia3D.Composition.Bushing — Type

force = Bushing(; obj1, obj2,
    nominalForce      = [0.0, 0.0, 0.0],
    springForceLaw    = [0.0, 0.0, 0.0],
    damperForceLaw    = [0.0, 0.0, 0.0],
    nominalTorque     = [0.0, 0.0, 0.0],
    rotSpringForceLaw = [0.0, 0.0, 0.0],
    rotDamperForceLaw = [0.0, 0.0, 0.0],
    largeAngles       = false )

Return a force acting as bushing between obj1::Object3D and obj2::Object3D. The force directions are defined by obj1, i.e. the orientation of obj2 does not influence the resulting forces.

Arguments

nominalForce defines the nominal force vector, i.e. the force that acts when spring and damper forces are zero. Positive values act in positive axis directions at obj1 and in opposite directions at obj2.

springForceLaw defines the force law of the spring in x-, y- and z-direction:

A Real number represents a linear stiffness coefficient.

An univariate Function is used to compute the spring force dependent of its deflection.

damperForceLaw defines the force law of the damper in x-, y- and z-direction:

A Real number represents a linear damping coefficient.

An univariate Function is used to compute the damper force dependent of its deflection velocity.

nominalTorque defines nominal torques about alpha, beta and gamma directions.

rotSpringForceLaw defines the force law of the rotational spring about alpha-, beta- and gamma-direction:

A Real number represents a linear damping coefficient.

An univariate Function is used to compute the spring force dependent of its deflection.

rotDamperForceLaw defines the force law of the rotational damper about alpha-, beta- and gamma-direction:

A Real number represents a linear damping coefficient.

An univariate Function is used to compute the damper force dependent of its deflection velocity.

largeAngles can be used to enable large angle mode.

When disabled, small deformation angles (< 10°) are assumed. This option deals equally with rotations [alpha, beta gamma] about the axes [x, y, z] of obj1, but causes approximation errors for larger angles.

When enabled, the deformation angles and torque directions are calculated as Cardan (Tait–Bryan) angles (rotation sequence x-y-z from obj1 to obj2). This option supports angles up to nearly 90°, but introduces local rotation directions [alpha, beta gamma] which differ from the axes [x, y, z] of obj1 and increases computation effort.

SpringDamperPtP

Modia3D.Composition.SpringDamperPtP — Type

force = SpringDamperPtP(; obj1, obj2,
    nominalLength  = 0.0,
    nominalForce   = 0.0,
    springForceLaw = 0.0,
    damperForceLaw = 0.0 )

Return a force acting as point-to-point parallel spring-damper between obj1::Object3D and obj2::Object3D.

Arguments

nominalLength defines the nominal length, i.e. the distance between obj1 and obj2 where the deflection of the spring is zero.

nominalForce defines the nominal force, i.e. the force that acts when spring and damper forces are zero. Positive values represent tension.

springForceLaw defines the force law of the spring:

A Real number represents a linear stiffness coefficient.

An univariate Function is used to compute the spring force dependent of its deflection. Positive values represent tension.

damperForceLaw defines the force law of the damper:

A Real number represents a linear damping coefficient.

An univariate Function is used to compute the damper force dependent of its deflection velocity. Positive values represent expansion.