Gravity Field

A global gravity field for the whole scene is defined with keyword gravitys in a Scene. Possible values:

NoGravityField

Modia3D.Composition.NoGravityField — Type

gravityField = NoGravityField()

Generate an instance of type NoGravityField that defines no gravity.

UniformGravityField

Modia3D.Composition.UniformGravityField — Type

UniformGravityField(;g=9.81, n=[0,-1,0])

Generate an instance of type UniformGravityField that defines a uniform gravity field with gravitational acceleration g in direction n.

Arguments

g::Float64: Gravitational acceleration

n::AbstractVector: Direction of gravitational acceleration

Example

import Modia3D

grav = Modia3D.Composition.UniformGravityField()
   r = Modia3D.EarthRadius
   g = gravityAcceleration(grav,r)  # g is independent of r

PointGravityField

Modia3D.Composition.PointGravityField — Type

PointGravityField([mass|;mue=G*EarthMass])

Generate an instance of type PointGravityField that defines a point gravity field of mass or gravity field constant mue. The gravity center is located at the world origin, i.e. the gravitational acceleration decreases quadratically with distance from world.

Example

import Modia3D

grav = Modia3D.PointGravityField()   # Gravity field of earth
   r = Modia3D.EarthRadius
   g = gravityAcceleration(grav,r)  # g is singular at r=0