Basics

Basics

Modia3D.Basics — Module.

module Modia3D.Basics

Utility constants and functions for Modia3D

Main developers

Andrea Neumayr and Martin Otter, DLR - Institute of System Dynamics and Control

Modia3D.Basics.BoundingBox — Type.

mutable struct BoundingBox - Smallest box that contains a visual element

Modia3D.Basics.PTP_path — Type.

path = PTP_path(names;
                positions = [zeros(size(names,1))'; ones(size(names,1))'],
                startTime=0.0,
                v_max = ones(size(names,1)),
                a_max = ones(size(names,2))))

Generate a new path object.

Modia3D.Basics.getAndCheckFullLibraryPath — Method.

getAndCheckFullLibraryPath(dir,libname)

Return joinpath(dir,libname). The returned full path must be a DLL. It is checked whether this DLL can be opened

Modia3D.Basics.getEnvironmentVariable — Method.

value = getEnvironmentVariable(name, description)

Returns the value of the environment variable name. If name does not exist, an error message is triggered:

Environment variable <name> not defined
(= <description>)

Modia3D.Basics.getIndex — Method.

index = getIndex(path, name)

Return the index of name in path or trigger an error, if not present.

Modia3D.Basics.getPosition! — Method.

getPosition!(path, time, position)

Given a path::PTP_path and a time instant time, return the actual position at time time in vector position.

Modia3D.Basics.getPosition — Method.

pos = getPosition(path, index, time)

Given a path::PTP_path, the index of a signal, and a time instant time, return the actual position at time time.

Modia3D.Basics.listKeys — Method.

listKeys(dict)

List the keys of dictionary dict in sorted order.

Modia3D.Basics.plotPath — Method.

plotPath(path;
         names=path.names, heading="PTP plots",
         tend=1.1*path.Tend, figure=1, ntime=101)

Given a path::PTP_path, plot the path over time up to tend for all points identified by the vector or tuple names to figure figure using ntime time points.

Modia3D.Basics.readDictOfStructsFromJSON — Method.

readDictOfStructsFromJSON(fileName, StructType)

Read a JSON file from fileName and return a Dict{String, StructType} dictionary. StructType must be a mutable struct type with a constructor StructType().