Composition

Composition

Modia3D.Composition — Module.

module Modia3D.Composition

Structuring of objects moving in 3D. Most important constructors (dof are the degrees-of-freedom):

Function	dof	Description

@assemblyName(..) begin .. end	-	Return a Name constructor for Object3Ds

Object3D([data];..)	0	Return a reference Object3D

Object3D(parent [, data];..)	0,6	Return Object3D fixed/moving w.r.t. parent

Modia3D.Revolute(obj1,obj2;..)	1	Return a revolute joint

Modia3D.Prismatic(obj1,obj2;..)	1	Return a prismatic joint

The optional data associated with an Object3D can be one of the following:

data	Description

::Modia3D.Solid	Solids with geometry, mass, visual/contact material

<:Modia3D.AbstractVisualElement	Visual elements (Modia3D.Graphics)

Main developers

Andrea Neumayr and Martin Otter, DLR - Institute of System Dynamics and Control

Modia3D.Composition.ContactDetectionMPR_handler — Type.

handler = ContactDetectionMPR_handler(;tol_rel = 1e-4, niter_max=100, neps=sqrt(eps()))

Generate a new contact handler for usage of the MPR algorithm The handler instance contains all information about the contact situation.

Arguments

tol_rel: Relative tolerance to compute the contact point (> 0.0)

niter_max: Maximum number of iterations of the MPR algorithm. If this number is reached, an error occurs (> 0).

neps: Small number used to check whether a floating number is close to zero (> 0.0).

Modia3D.Composition.ContactPair — Type.

pair = ContactPair(contactPoint1,contactPoint2,contactNormal,obj1,obj2,
                   distanceWithHysteresis)

Generate a new ContactPair object of two objects that have contact=true.

Modia3D.Composition.ElasticContactPairResponseMaterial — Type.

m = ElasticContactPairResponseMaterial(c_res, d_res, mu_k, mu_r, vsmall, wsmall)

Return an ElasticContactPairResponseMaterial < AbstractContactPairMaterial object.

Modia3D.Composition.NoContactPair — Type.

pair = NoContactPair(contactPoint1,contactPoint2,contactNormal,obj1,obj2,
                     distanceWithHysteresis)

Generate a new NoContactPair object of two objects that have contact=false.

Modia3D.Composition.NoGravityField — Type.

gravityField = NoGravityField()

Generate an instance of type NoGravityField that defines no gravity.

Modia3D.Composition.Object3D — Type.

obj1 = Object3D([data]; visualizeFrame=Modia3D.Inherited)
obj2 = Object3D(parent [, data]; fixed=true, r=zeros(3), R=nothing, q=nothing,
                v_start=zeros(3), w_start=zeros(3),
                visualizeFrame=Modia3D.Inherited)

Generate a new Object3D object, that is a coordinate system (= frame) with associated data. If parent is present, the Object3D is defined relatively to the parent Object3D. If parent is not present, the Object3D is either a reference object (such as the world-object), or the object is connected later with a joint to another Object3D. If fixed=true, the object is rigidly connect to its parent; otherwise it is moving freely relative to its parent (mathematically described by quaternions).

Note, there are many convenience functions in ModiaMath.Frames to generate a ModiaMath.RotationMatrix R or a ModiaMath.Quaternion q.

Arguments

data::Modia3D.AbstractObject3Ddata: Optional data associated with Object3D.

parent::Object3D: Parent object.

fixed::Bool:

fixed = true, if the Object3D is fixed relatively to its parent Object3D at position r,R,q. It is best to provide the rotation information via R in this case.

fixed = false, if Object3D can move freely relatively to its parent Object3D and is initially placed at r,R,q. The movement is internally described with Quaternion vector q. Therefore, it is best to provide the rotation information via q in this case.

r::AbstractVector: Initial relative position vector from frame of parent object to origin of frame object, resolved in parent frame.

R::Union{ModiaMath.RotationMatrix,NOTHING}: Initial rotation matrix defining the rotation from frame of parent object to frame of Object3D. If both R = nothing and q = nothing, a null rotation is defined.

q::Union{ModiaMath.Quaternion,NOTHING}: Initial quaternion defining the rotation from frame of parent object to frame of Object3D. If both R = nothing and q = nothing, a null rotation is defined.

v_start::AbstractVector: If fixed=false, initial velocity of the origin of Object3D with respect to parent, resolved in parent frame.

w_start::AbstractVector: If fixed=false, initial angular velocity of Object3D with respect to parent, resolved in Object3D.

visualizeFrame::Union{Bool,Modia3D.Ternary}: Coordinate system of Object3D is always (= true), or never (= false) visualized, or it is visualized if defined in SceneOptions(...) (= Modia3D.Inherited).

Examples

using Modia3D

# Define assembly
@assembly MyAssembly begin
   world = Object3D()

   # Frame fixed in world
   frame1 = Object3D(world; r=[0.1, 0.2, 0.3])

   # Frame moving relatively to frame1
   r2     = [0.2, 0.2, 0.3]
   frame2 = Object3D(frame1; fixed=false, r=r2)

   # Frame moving relatively to world
   frame3 = Object3D(world; fixed=false, r=-r2)
end
Modia3D.visualizeAssembly!(MyAssembly())

Modia3D.Composition.PointGravityField — Type.

PointGravityField([mass|;mue=G*EarthMass])

Return a PointGravityField struct with the gravity field constant mue (mue = G*mass).

Example

import Modia3D

grav = Modia3D.PointGravityField()   # Gravity field of earth
   r = Modia3D.EarthRadius
   g = gravityAcceleration(grav,r)

Modia3D.Composition.SceneOptions — Type.

sceneOptions = Modia3D.SceneOptions(;kwargs...)

Define global options for a simulation of the scene with keyword arguments:

Keyword arguments	defaults

enableContactDetection	true

elasticContactReductionFactor	1.0

gravityField	Modia3D.UniformGravityField()

enableVisualization	true

visualizeGravity	true

visualizeFrames	false

visualizeConvexHulls	true

visualizeContactPoints	false

visualizeSupportPoints	false

nominalLength	1.0

defaultFrameLength	0.2*nominalLength

defaultJointLength	nominalLength/10

defaultJointWidth	nominalLength/20

defaultForceLength	nominalLength/10

defaultForceWidth	nominalLength/20

defaultBodyDiameter	nominalLength/9

defaultWidthFraction	20

defaultArrowDiameter	nominalLength/40

defaultN_to_m	1000

defaultNm_to_m	1000

useOptimizedStructure	true

defaultContactSphereDiameter	0.1

contactDetection	ContactDetectionMPR_handler()

Optional keyword arguments

enableContactDetection::Bool: = true, if contact detection is enabled

elasticContactReductionFactor::Float64: usedcontactcompliance = contact_compliance * elasticContactReductionFactor (> 0).

gravityField::Modia3D.AbstractGravityField: Type of gravity field, such as: Modia3D.NoGravityField, Modia3D.UniformGravityField, Modia3D.PointGravityField.

enableVisualization::Bool: = true, if animation is enabled

visualizeGravity::Bool: = true, if gravity field shall be visualized (acceleration vector or field center)

visualizeFrames::Bool: = true, if all frames shall be visualized

visualizeConvexHulls::Bool: = true, if convex hulls (used for contact detection) shall be visualized

visualizeContactPoints::Bool: = true, if contact points shall be visualized

visualizeSupportPoints::Bool = true, if support points shall be visualized

nominalLength::Float64: [m] Nominal length of 3D system

defaultFrameLength::Float64: [m] Default for frame length if visualizeFrames = true (but not world frame)

defaultJointLength::Float64: [m] Default for the fixed length of a shape representing a joint

defaultJointWidth::Float64: [m] Default for the fixed width of a shape representing a joint

defaultForceLength::Float64: [m] Default for the fixed length of a shape representing a force (e.g., damper)

defaultForceWidth::Float64: [m] Default for the fixed width of a shape representing a force (e.g., spring, bushing)

defaultBodyDiameter::Float64: [m] Default for diameter of sphere representing the center of mass of a body

defaultWidthFraction::Float64: Default for shape width as a fraction of shape length

defaultArrowDiameter::Float64: [m] Default for arrow diameter (e.g., of forces, torques, sensors)

defaultN_to_m::Float64: [N/m] Default scaling of force arrows (length = force/defaultNtom)

defaultNm_to_m::Float64: [N.m/m] Default scaling of torque arrows (length = torque/defaultNmtom)

useOptimizedStructure::Bool: = true, if the optimized structure (with super objects, and common inertia) is used

defaultContactSphereDiameter::Float64: [m] Diameter of sphere used for contact point visualization

contactDetection::Modia3D.AbstractContactDetection: Handler used for contact detection (for example to determine the smallest distance between two objects).

Example

For all the details see "Modia3D/examples/dynamics/Simulate_Pendulum.jl":

@assembly Pendulum(;Lx = 1.0) begin
   ...
end

pendulum = Pendulum(Lx=1.6, sceneOptions=
             Modia3D.SceneOptions(visualizeFrames=true, defaultFrameLength=0.3))
model    = Modia3D.SimulationModel( pendulum )
result   = ModiaMath.simulate!(model, stopTime=4.5)

Modia3D.Composition.SimulationModel — Type.

simModel = SimulationModel(assembly::Modia3D.AbstractAssembly;
                           analysis::ModiaMath.AnalysisType=ModiaMath.DynamicAnalysis,
                           startTime = 0.0, stopTime  = 1.0, tolerance = 1e-4,
                           interval  = (stopTime-startTime)/500.0,
                           maxStepSize = NaN, maxNumberOfSteps=missing)

Generate a simulationModel from an assembly generated with macro Modia3D.@assembly and the type of analysis to be carried out on the assembly. Additionally, default startTime, stopTime, tolerance, interval, maxStepSize, maxNumberOfSteps, for the simulation engine are defined. These values should be adapted so that assembly-specific, meaningful defaults are provided.

Modia3D.Composition.UniformGravityField — Type.

UniformGravityField(;g=9.81, n=[0,1,0])

Return a UniformGravityField struct.

Arguments

g::Float64: Gravity constant

n::AbstractVector: Direction of gravity

Example

import Modia3D

grav = Modia3D.UniformGravityField()
   r = Modia3D.EarthRadius
   g = gravityAcceleration(grav,r)

Modia3D.Composition.Fixed — Method.

joint = Modia3D.Fixed(obj1, obj2; r=zeros(3), R=nothing, q=nothing)

Return a Fixed joint that fixes obj2::Object3D relatively to obj1::Object3D via the relative position vector from obj1 to obj2 (resolved in obj1) and the relative transformation matrix R from obj1 to obj2 or alternatively the relative quaternion q.

Arguments

obj1::Object3D: Parent object.

obj2::Object3D: Object fixed relative to obj1.

r::AbstractVector: Relative position vector from obj1 to obj2 resolved in obj1.

R::Union{ModiaMath.RotationMatrix,Nothing}: Rotation matrix defining the rotation from obj1 to obj. If both R = nothing and q = nothing, a null rotation is defined.

q::Union{ModiaMath.Quaternion,Nothing}: Quaternion defining the rotation from obj1 to obj2. If both R = nothing and q = nothing, a null rotation is defined.

Modia3D.Composition.Prismatic — Method.

joint = Modia3D.Prismatic(obj1::Object3D, obj2::Object3D;
                          axis=1, s_start=0, v_start=0, canCollide=false)

Return a joint object that constrains the movement of obj2::Object3D with respect to obj1::Object3D along coordinate axis axis (axis = 1,2,3,-1,-2,-3). The initial position/velocity of obj2 with respect to obj1 along axis is s_start [m] and v_start [m/s], respectively. If canCollide=false, no collision detection will occur between obj1 and obj2 (and Object3Ds that are directly or indirectly rigidly fixed to obj1 or obj2).

If a Prismatic joint closes a kinematic loop, then the already present objects must be consistent to the Prismatic joint that is the frames of obj1 and obj2 must be parallel to each other and movement of obj1 along its axis axis with s_start results in obj2. If s_start=NaN, its value is computed in this case.

Examples

using Modia3D
import ModiaMath

@assembly FallingBall(;h=1.0) begin
   world  = Object3D()
   sphere = Object3D( Modia3D.Solid(Modia3D.SolidSphere(0.1), "Aluminium") )

   # Constrain sphere movement (initial placement at position [0,h,0])
   prismatic = Modia3D.Prismatic(world, sphere, axis=2, s_start=h)
end

simulationModel = Modia3D.SimulationModel( FallingBall(h=1.5), stopTime=1.0 )
result          = ModiaMath.simulate!(simulationModel)
ModiaMath.plot(result, ("prismatic.s", "prismatic.v"))

Modia3D.Composition.Revolute — Method.

joint = Modia3D.Revolute(obj1, obj2; axis=3, phi_start=0, w_start=0, canCollide=false, J=NaN)

Return a Revolute joint that rotates obj1::Object3D into obj2::Object3D along the axis axis of obj1 (axis = 1,2,3,-1,-2,-3). The initial start angle is phi_start and the initial angular velocity is w_start. If canCollide=false, no collision detection will occur between obj1 and obj2 (and Object3Ds that are directly or indirectly rigidly fixed to obj1 or obj2).

Optionally, an inertia can be placed on the axis of rotation (J). If such an inertia is provided, the generalized coordinate is still the angle of the revolute joint and the driving torque is still the torque acting in the revolute joint. If, for example, an electric motor with a gearbox shall be modelled then

J = JMotor*gearRatio^2
jointTorque = motorTorque*gearRatio
motor

``J = JMotor*gearRatio^2; jointTorque =.

If a Revolute joint closes a kinematic loop, then the already present objects must be consistent to the Revolute joint. That is, the frames of obj1 and obj2 must be parallel to each other and rotating obj1 along its axis axis with phi_start results in obj2. If phi_start=NaN, its value is computed in this case.

Examples

using Modia3D
import ModiaMath

@assembly Pendulum(;L=1.0) begin
   world  = Modia3D.Object3D()
   obj1 = Modia3D.Object3D()
   rev    = Modia3D.Revolute(world, obj1)
   sphere = Modia3D.Object3D(obj1, Modia3D.Solid(Modia3D.SolidSphere(0.1), "Aluminium"),
                             r = [L,0,0] )
end

simulationModel = Modia3D.SimulationModel( Pendulum(L=0.8), stopTime=5.0 )
result          = ModiaMath.simulate!(simulationModel)
ModiaMath.plot(result, ("rev.phi", "rev.w"))

Modia3D.Composition.contactStart — Method.

resonseMaterial = contactStart(matPair::Modia3D.ElasticContactPairMaterial, obj1,obj2,rContact,contactNormal)

Return a responseMaterial::ElasticContactPairResponseMaterial object at the start of a collision.

Modia3D.Composition.contactStart — Method.

responseMaterial = contactStart(obj1, obj2, rContact, contactNormal)

Return a response material object at contact start.

Modia3D.Composition.distance — Method.

d = distance(frame1, frame2)

Return the distance between the origin of frame1 and the origin of frame2

Modia3D.Composition.normalRelativeVelocityAtContact — Method.

delta_dot = normalRelativeVelocityAtContact(obj1, obj2, rContact, contactNormal)

Return the relative velocity in normal direction contactNormal at contact point rContact of the two penetrating objects obj1, obj2.

Modia3D.Composition.planarRotationAngle — Method.

angle = planarRotationAngle(e, v1, v2)
      = planarRotationAngle(frame1, frame2)

Return angle of a planar rotation, given the rotation axis e (a unit vector) and the representations of a vector in frame 1 (v1) and frame 2 (v2).

Under the assumption that the z-axes of frame1 and frame2 coincide, return the angle between the x-axis of frame1 and the position vector from frame1 to frame2.

Modia3D.Composition.responseCalculation — Method.

(f1,f2,t1,t2) = responseCalculation(material::ElasticContactPairResponseMaterial,
                        obj1, obj2, rContact, contactNormal, distanceWithHysteresis, time)

Compute contact forces f1,f2 and contact torques t1,t2 from the elastic contact pair response material material the two penetrating objects obj1, obj2, the contact point rContact, the contact normal contactNormal and the largest penetration depth distanceWithHysteresis at time time.

Modia3D.Composition.visualizeAssembly! — Method.

visualizeAssembly!(assembly::Modia3D.AbstractAssembly)

Visualize the assembly defined with macro Modia3D.@assembly in its initial configuration (but without simulating it).

Modia3D.Composition.@assembly — Macro.

@assembly AssemblyName(arguments) begin ... end

Return the constructor for a new struct AssemblyName consisting of Object3Ds that are connected together. The new struct consists of all left-hand-side (scalar or vector) symbols present between begin ... end.

Examples

using Modia3D

@assembly Bar(;Lx = 0.1, Ly=Lx/5, Lz=Ly) begin
   frame0 = Modia3D.Object3D(Modia3D.Solid(Modia3D.SolidBeam(Lx,Ly,Lz), nothing, vmat1))
   frame1 = Modia3D.Object3D(frame0; r=[-Lx/2, 0.0, 0.0])
   frame2 = Modia3D.Object3D(frame0; r=[ Lx/2, 0.0, 0.0])
   cyl1   = Modia3D.Object3D(frame1, cyl)
   cyl2   = Modia3D.Object3D(frame2, cyl)
end
bar = Bar(;Lx=1.0)
Modia3D.visualizeAssembly!( bar )