ForceElements

ForceElements

Modia3D.ForceElements — Module.

module Modia3D.ForceElements

Elements generating forces or torques that act on objects.

Main developers

Andrea Neumayr and Martin Otter, DLR - Institute of System Dynamics and Control

source

Modia3D.ForceElements.@forceElement — Macro.

@forceElement forceName(arguments) begin ... end - Generate a function that instantiates a new forceElement

forceName(arguments) begin ... end is treated as the constructor function of a new (mutable) struct that consists of all left-hand-side (scalar or vector) symbols present between begin ... end.

Examples

julia> include("D:\otter\_github\Modia3D.jl/examples/dynamics/Simulate_PendulumWithDamper.jl");

using ModiaMath
using Modia3D

@forceElement Damper(; d=1.0) begin
    w   = ModiaMath.RealScalar("w",   causality=ModiaMath.Input,  numericType=ModiaMath.WR)
    tau = ModiaMath.RealScalar("tau", causality=ModiaMath.Output, numericType=ModiaMath.WR)
end;
function Modia3D.computeTorque(damper::Damper, sim::ModiaMath.SimulationState)
    damper.tau.value = -damper.d*damper.w.value
end;
@assembly PendulumWithDamper(;Lx = 1.0, Ly=0.2*Lx, Lz=0.2*Lx, m=1.0, g=9.81) begin
   world  = Modia3D.Object3D(Modia3D.CoordinateSystem(0.5*Lx))
   body   = Modia3D.Object3D(Modia3D.Solid(Modia3D.SolidBeam(Lx,Ly,Lz), Modia3D.MassProperties(m=m), vmat1))
   frame1 = Modia3D.Object3D(body; r=[-Lx/2, 0.0, 0.0])
   cyl    = Modia3D.Object3D(frame1,Modia3D.Cylinder(Ly/2,1.2*Ly; material=vmat2))

   # Connect pendulum to world with a damper in the joint
   rev    = Modia3D.Revolute(world, frame1)
   d      = Damper(d=0.2)
   damper = Modia3D.AdaptorForceElementToFlange(w=d.w, tau=d.tau)
   Modia3D.connect(damper, rev)
end
pendulum = PendulumWithDamper()
model = Modia3D.SimulationModel( pendulum )
result = ModiaMath.simulate!(model, stopTime=5.0, interval=0.1, tolerance=1e-4, log=true)
ModiaMath.plot(result, ["rev.phi", "rev.w", "rev.a", "rev.tau"])

source